update the clone script to handle non-regular repositories

+++ clone.sh
@@ -4,7 +4,6 @@ WORKING_DIR=`pwd`
 REPOS="$WORKING_DIR/repos"
 REPO_LIST="$WORKING_DIR/repo-list.json"
 
-
 cd $REPOS
 
 let SIZE=$(cat $REPO_LIST | jq -r '.repos | length')-1
@@ -15,16 +14,24 @@ do
 	echo "Repo: $NAME"
 	URL=$(cat $REPO_LIST | jq -r ".repos | .[$i].url")
 	COMMIT=$(cat $REPO_LIST | jq -r ".repos | .[$i].commit")
-	if [ ! -d "$NAME" ]; then
-		echo "$NAME does not exist."
-		git clone $URL
+	REGULAR=$(cat $REPO_LIST | jq -r ".repos |.[$i].status.regular")
+	if [ "$REGULAR" = true ]; then
 		if [ ! -d "$NAME" ]; then
-			echo "git failed to clone $REPO"
-		else
-			cd $NAME
-			git reset --hard $COMMIT
-			cd $REPOS
+			echo "$NAME does not exist."
+			git clone $URL $NAME
+			if [ ! -d "$NAME" ]; then
+				echo "git failed to clone $REPO"
+			else
+				cd $NAME
+				git reset --hard $COMMIT
+				cd $REPOS
+			fi
 		fi
+	else
+		echo "non regular"
+		cd $WORKING_DIR
+		config/$NAME/build.sh --name $NAME --url $URL --commit $COMMIT
+		cd $REPOS
 	fi
 done
